pinctrl: msm: Remove explicit barriers from mmio ops where unneeded
For the vast majority of mmio operations in this driver, explicit memory
barriers aren't needed either because a data dependency between a read
and write already exists, or because of the presence of the spin locks
which execute a full memory barrier.

Removing all the unneeded explicit barriers considerably reduces
overhead for pinctrl operations, which in turn benefits things like i2c.

Change-Id: I23d9d3d12ca3d2648f2d0f96f5f7d0c46d9ead09
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: dodyirawan85 <40514988+dodyirawan85@users.noreply.github.com>